% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornell_colors.R
\name{render_cornell_colors}
\alias{render_cornell_colors}
\title{Cornell colors palette generator.}
\usage{
render_cornell_colors(palette_name, n)
}
\arguments{
\item{palette_name}{Name of desired palette. Choices are:
\code{classic}, \code{secondary}, \code{accents}, \code{malott}, \code{brb}}

\item{n}{Number of colors desired. Most palettes have 3-5 colors. Most color
schemes are derived from \href{https://brand.cornell.edu/design-center/colors/}.
If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
Cornell colors palette generator.
}
\examples{
render_cornell_colors("classic")
render_cornell_colors("malott", 3)
}
\keyword{colors}
